+++ Api_LigaDeportiva/Controllers/PartidosController.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Api_LigaDeportiva.Data;
@@ -23,33 +22,59 @@ namespace Api_LigaDeportiva.Controllers
 
         // GET: api/Partidos
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Partido>>> GetPartido()
+        public async Task<ActionResult<ApiResult<List<Partido>>>> GetPartido()
         {
-            return await _context.Partido.ToListAsync();
+            try
+            {
+                var data = await _context.Partido
+                    .Include(p => p.Torneo)
+                    .Include(p => p.Resultado)
+                    .ToListAsync();
+
+                return ApiResult<List<Partido>>.Ok(data);
+            }
+            catch (Exception ex)
+            {
+                return ApiResult<List<Partido>>.Fail(ex.Message);
+            }
         }
 
         // GET: api/Partidos/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Partido>> GetPartido(int id)
+        public async Task<ActionResult<ApiResult<Partido>>> GetPartido(int id)
         {
-            var partido = await _context.Partido.FindAsync(id);
-
-            if (partido == null)
+            try
             {
-                return NotFound();
-            }
+                var partido = await _context.Partido
+                    .Include(p => p.Torneo)
+                    .Include(p => p.Resultado)
+                    .FirstOrDefaultAsync(p => p.Id == id);
 
-            return partido;
+                if (partido == null)
+                {
+                    return ApiResult<Partido>.Fail("Partido no encontrado.");
+                }
+
+                return ApiResult<Partido>.Ok(partido);
+            }
+            catch (Exception ex)
+            {
+                return ApiResult<Partido>.Fail(ex.Message);
+            }
         }
 
-        // PUT: api/Partidos/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutPartido(int id, Partido partido)
+        public async Task<ActionResult<ApiResult<Partido>>> PutPartido(int id, Partido partido)
         {
             if (id != partido.Id)
             {
-                return BadRequest();
+                return ApiResult<Partido>.Fail("No coinciden los identificadores.");
+            }
+
+            var yaJugado = await _context.Resultado.AnyAsync(r => r.PartidoId == id);
+            if (yaJugado)
+            {
+                return ApiResult<Partido>.Fail("El partido ya fue jugado y sus resultados han sido registrados. No se puede modificar el calendario.");
             }
 
             _context.Entry(partido).State = EntityState.Modified;
@@ -58,46 +83,70 @@ namespace Api_LigaDeportiva.Controllers
             {
                 await _context.SaveChangesAsync();
             }
-            catch (DbUpdateConcurrencyException)
+            catch (DbUpdateConcurrencyException ex)
             {
                 if (!PartidoExists(id))
                 {
-                    return NotFound();
+                    return ApiResult<Partido>.Fail("Datos no encontrados.");
                 }
                 else
                 {
-                    throw;
+                    return ApiResult<Partido>.Fail(ex.Message);
                 }
             }
 
-            return NoContent();
+            return ApiResult<Partido>.Ok(partido);
         }
 
         // POST: api/Partidos
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPost]
-        public async Task<ActionResult<Partido>> PostPartido(Partido partido)
+        public async Task<ActionResult<ApiResult<Partido>>> PostPartido(Partido partido)
         {
-            _context.Partido.Add(partido);
-            await _context.SaveChangesAsync();
+            try
+            {
+                if (!await _context.Torneo.AnyAsync(t => t.Id == partido.TorneoId))
+                {
+                    return ApiResult<Partido>.Fail("El Torneo especificado no existe.");
+                }
 
-            return CreatedAtAction("GetPartido", new { id = partido.Id }, partido);
+                _context.Partido.Add(partido);
+                await _context.SaveChangesAsync();
+
+                return ApiResult<Partido>.Ok(partido);
+            }
+            catch (Exception ex)
+            {
+                return ApiResult<Partido>.Fail(ex.Message);
+            }
         }
 
         // DELETE: api/Partidos/5
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeletePartido(int id)
+        public async Task<ActionResult<ApiResult<Partido>>> DeletePartido(int id)
         {
-            var partido = await _context.Partido.FindAsync(id);
-            if (partido == null)
+            try
             {
-                return NotFound();
+                var partido = await _context.Partido.FindAsync(id);
+                if (partido == null)
+                {
+                    return ApiResult<Partido>.Fail("Partido no encontrado.");
+                }
+
+                var yaJugado = await _context.Resultado.AnyAsync(r => r.PartidoId == id);
+                if (yaJugado)
+                {
+                    return ApiResult<Partido>.Fail("No se puede eliminar un partido que ya tiene resultados registrados.");
+                }
+
+                _context.Partido.Remove(partido);
+                await _context.SaveChangesAsync();
+
+                return ApiResult<Partido>.Ok(null);
             }
-
-            _context.Partido.Remove(partido);
-            await _context.SaveChangesAsync();
-
-            return NoContent();
+            catch (Exception ex)
+            {
+                return ApiResult<Partido>.Fail(ex.Message);
+            }
         }
 
         private bool PartidoExists(int id)
